+++ test/cli/install/bun-link.test.ts
@@ -1,9 +1,9 @@
 // @known-failing-on-windows: 1 failing
 import { spawn, file } from "bun";
 import { afterAll, afterEach, beforeAll, beforeEach, expect, it } from "bun:test";
-import { bunExe, bunEnv as env } from "harness";
+import { bunExe, bunEnv as env, toBeValidBin, toHaveBins } from "harness";
 import { access, mkdtemp, readlink, realpath, rm, writeFile, mkdir } from "fs/promises";
-import { basename, join } from "path";
+import  { basename, join, sep, dirname } from "path";
 import { tmpdir } from "os";
 import {
   dummyAfterAll,
@@ -18,6 +18,11 @@ beforeAll(dummyBeforeAll);
 afterAll(dummyAfterAll);
 
 let link_dir: string;
+
+expect.extend({
+  toBeValidBin,
+  toHaveBins,
+})
 
 beforeEach(async () => {
   link_dir = await mkdtemp(join(await realpath(tmpdir()), "bun-link.test"));
@@ -68,8 +73,8 @@ it("should link and unlink workspace package", async () => {
   var out = await new Response(stdout).text();
   expect(out.replace(/\s*\[[0-9\.]+ms\]\s*$/, "").split(/\r?\n/)).toEqual([
     "",
-    " + boba@workspace:packages/boba",
-    " + moo@workspace:packages/moo",
+    ` + boba@workspace:packages${sep}boba`,
+    ` + moo@workspace:packages${sep}moo`,
     "",
     " 2 packages installed",
   ]);
@@ -393,7 +398,6 @@ it("should link scoped package", async () => {
 });
 
 it("should link dependency without crashing", async () => {
-  console.log(link_dir);
   const link_name = basename(link_dir).slice("bun-link.".length) + "-really-long-name";
   await writeFile(
     join(link_dir, "package.json"),
@@ -461,8 +465,8 @@ it("should link dependency without crashing", async () => {
   ]);
   expect(await exited2).toBe(0);
   expect(await readdirSorted(join(package_dir, "node_modules"))).toEqual([".bin", ".cache", link_name].sort());
-  expect(await readdirSorted(join(package_dir, "node_modules", ".bin"))).toEqual([link_name]);
-  expect(await readlink(join(package_dir, "node_modules", ".bin", link_name))).toBe(
+  expect(await readdirSorted(join(package_dir, "node_modules", ".bin"))).toHaveBins([link_name]);
+  expect(join(package_dir, "node_modules", ".bin", link_name)).toBeValidBin(
     join("..", link_name, `${link_name}.js`),
   );
   expect(await readdirSorted(join(package_dir, "node_modules", link_name))).toEqual(
